--- ScratchScript/src/ScratchScript/function_specs/hashes.json
@@ -1,1 +1,0 @@
-{"Looks":"2272f6c562920538ad0485d150c3a6bdbc398678bb457bdd5b81ba16ead717b9","Motion":"b908ed0fa335ea50343fbc25f647cc0c2abe9b3f094f954f98967f548503903b","Variables":"4af2c9e7584b8b6864effbe637930f4a2a18c320dabebcb26afa9911f72034ce"}
