--- ScratchScript/src/ScratchScript/lang_compiler/__init__.py
@@ -1,6 +1,0 @@
-from ScratchScript.lang_compiler.lang_extract import parse_program
-
-
-def main(program):
-    program = parse_program(program)
-
